[ADD]新增了POIUtil类导出excel表格。 [ADD]在devController中增加了方法以演示如何导入excel表格

+++ src/main/java/com/gdou/teaching/dto/UserDTO.java
@@ -16,7 +16,7 @@ public class UserDTO {
     private String userNumber;
     private String headUrl;
     private Integer classId;
-    private Integer userStatus;
+    private Byte userStatus;
     private Integer userIdent;
     private String mail;
     private String phone;
